requirements.txt
RPi.GPIO
flask
graypy
requests
